商品管理中实现商品模型和商品属性模块

+++ shop-manager/src/main/java/com/xxxx/manager/pojo/GoodsAttributeExample.java
@@ -258,7 +258,7 @@ public class GoodsAttributeExample {
             return (Criteria) this;
         }
 
-        public Criteria andTypeIdEqualTo(Short value) {
+        public Criteria andTypeIdEqualTo(Integer value) {
             addCriterion("type_id =", value, "typeId");
             return (Criteria) this;
         }
